finalized implementation of deleting invoices and invoice items, finished api documentation for all routes, modified return format for delete route on companies controller

+++ src/invoice-items/invoice-items.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Error } from 'src/common/enums/error.enum';
 import { Repository } from 'typeorm';
 import { CreateInvoiceItemDto } from './dto/create-invoice-item.dto';
 import { UpdateInvoiceItemDto } from './dto/update-invoice-item.dto';
@@ -42,8 +43,33 @@ export class InvoiceItemsService {
     return await this.invoiceItemsRepository.save(invoiceItem);
   }
 
-  async removeInvoiceItem(id: InvoiceItemId): Promise<void> {
+  // Hard delete
+  async deleteInvoiceItem(id: InvoiceItemId): Promise<void> {
     const invoiceItem = await this.getInvoiceItem(id);
+    if (!invoiceItem) {
+      throw new NotFoundException(Error.INVOICE_ITEM_NOT_FOUND);
+    }
     await this.invoiceItemsRepository.remove(invoiceItem);
+  }
+
+  // Soft delete
+  async removeInvoiceItem(id: InvoiceItemId): Promise<InvoiceItemId | null> {
+    const invoiceItem = await this.getInvoiceItem(id);
+    if (!invoiceItem) {
+      throw new NotFoundException(Error.INVOICE_ITEM_NOT_FOUND);
+    }
+    invoiceItem.archived = true;
+    await this.invoiceItemsRepository.save(invoiceItem);
+    return invoiceItem.id;
+  }
+
+  async restoreInvoiceItem(id: InvoiceItemId): Promise<InvoiceItemId | null> {
+    const invoiceItem = await this.getInvoiceItem(id);
+    if (!invoiceItem) {
+      throw new NotFoundException(Error.INVOICE_ITEM_NOT_FOUND);
+    }
+    invoiceItem.archived = false;
+    await this.invoiceItemsRepository.save(invoiceItem);
+    return invoiceItem.id;
   }
 }
